refactor(app): split game_service.dart by concern (Refs #2575)
Move turn-trace export and new-game setup pipelines into part files so
game_service.dart drops below the 1000 NCL threshold while preserving
the public GameService API unchanged.
